Move IntPreference to a separate file
